[refactor] change directory
